api/requirements.txt
amqp==5.1.1
Arpeggio==2.0.0
asgiref==3.5.2
attrs==22.1.0
azure-core==1.26.0
azure-storage-blob==12.13.1
azure-identity==1.11.0
billiard==3.6.4.0
cachetools==5.2.0
cairocffi==1.3.0
CairoSVG==2.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
cryptography==37.0.4
cssselect2==0.6.0
cycler==0.11.0
# datapunt-authorization-django==1.3.2
deepdiff
debugpy==1.6.3
debtcollector==2.5.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.5
Django==3.2.16
django-appconf==1.0.5
django-celery-beat==2.3.0
django-celery-email==3.0.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
django-markdownx==3.0.1
django-rest-swagger==2.2.0
django-storage-swift==1.2.19
django-storages==1.12.3
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
djangorestframework-xml==2.0.0
drf-amsterdam==0.4.1
drf-extensions==0.7.1
elasticsearch==6.8.2
elasticsearch-dsl==6.4.0
execnet==1.9.0
factory-boy==3.2.1
Faker==13.15.1
filelock==3.7.1
flake8==5.0.4
flex==6.14.1
fonttools==4.34.4
freezegun==1.2.1
future==0.18.2
google-api-core==2.10.1
google-auth==2.12.0
googleapis-common-protos==1.56.4
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
iso8601==1.0.2
isodate==0.6.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.9.1
jwcrypto==1.4.2
keystoneauth1==5.0.0
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mistune==2.0.4
msal==1.20.0
msal-extensions==1.0.0
msgpack==1.0.4
msrest==0.7.1
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.1
openapi-codec==1.3.2
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
os-service-types==1.7.0
oslo.config==9.0.0
oslo.i18n==5.1.0
oslo.serialization==5.0.0
oslo.utils==6.0.0
packaging==21.3
pbr==5.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.5.1
prompt-toolkit==3.0.30
protobuf==4.21.7
psutil==5.9.2
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
pyparsing==3.0.9
pyphen==0.12.0
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-timeout==2.1.0
pytest-xdist==2.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-keystoneclient==5.0.0
python-magic==0.4.27
python-swiftclient==4.0.1
pytz==2022.1
PyYAML==6.0
raven==6.10.0
requests==2.28.1
requests-mock==1.9.3
requests-oauthlib==1.3.1
rfc3986==2.0.0
rfc3987==1.3.8
rsa==4.9
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
strict-rfc3339==0.7
textX==3.0.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tox==3.25.1
types-cryptography==3.3.23
typing_extensions==4.4.0
tzdata==2022.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.11
uWSGI==2.0.20
validate-email==1.3
vine==5.0.0
virtualenv==20.16.3
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
wrapt==1.14.1
xmlunittest==0.5.0
zipp==3.8.1
exifread
piexif
GPSPhoto
